validacao

+++ src/persistencia/PCentroCusto.java
@@ -13,18 +13,19 @@ import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
+import javax.swing.JOptionPane;
 
 /**
  *
  * @author diogo
  */
 public class PCentroCusto {
-    
-     public void incluir(CentroCusto parametro) throws SQLException {
+
+    public void incluir(CentroCusto parametro) throws SQLException {
 
         String sql = " INSERT INTO "
-                   + " centro_custo (descricao, status ) "
-                   + " VALUES (?,?) ";
+                + " centro_custo (descricao, status ) "
+                + " VALUES (?,?) ";
 
         Connection conexao = util.Conexao.getConexao();
 
@@ -32,7 +33,7 @@ public class PCentroCusto {
 
         prd.setString(1, parametro.getDescricao());
         prd.setInt(2, parametro.getStatus());
-        
+
         prd.execute();
         conexao.close();
     }
@@ -40,8 +41,8 @@ public class PCentroCusto {
     public void alterar(CentroCusto parametro) throws SQLException {
 
         String sql = " UPDATE centro_custo "
-                   + " SET descricao = ?, status = ? "
-                   + " WHERE id = ?;";
+                + " SET descricao = ?, status = ? "
+                + " WHERE id = ?;";
 
         Connection conexao = util.Conexao.getConexao();
 
@@ -56,21 +57,37 @@ public class PCentroCusto {
     }
 
     public void excluir(int parametro) throws SQLException {
+        try {
 
-        String sql = " DELETE FROM centro_custo "
+            //verificar se o fornecedor possui título
+            PTitulosPagar ptp = new PTitulosPagar();
+            boolean fTitulo = ptp.ccTitulo(parametro);
+            if (fTitulo) {
+                throw new Exception("Centro de Custo com título vinculado \nNão é possível excluir apenas desativar!");
+            }
+            
+            /////////////////////////////////////////
+
+            String sql = " DELETE FROM centro_custo "
                     + " WHERE id = ?;";
 
-        Connection conexao = util.Conexao.getConexao();
-        PreparedStatement prd = conexao.prepareStatement(sql);
-        prd.setInt(1, parametro);
-        prd.execute();
-        conexao.close();
+            Connection conexao = util.Conexao.getConexao();
+            PreparedStatement prd = conexao.prepareStatement(sql);
+            prd.setInt(1, parametro);
+            prd.execute();
+            conexao.close();
+
+            JOptionPane.showMessageDialog(null, "Exclusão efetuado com Sucesso");
+         
+        } catch (Exception e) {
+            JOptionPane.showMessageDialog(null, e.getMessage());
+        }
     }
 
     public CentroCusto consultar(int parametro) throws SQLException {
 
         String sql = " SELECT id, descricao, status FROM centro_custo"
-                   + " WHERE id = ?;";
+                + " WHERE id = ?;";
 
         Connection conexao = util.Conexao.getConexao();
         PreparedStatement prd = conexao.prepareStatement(sql);
@@ -84,14 +101,14 @@ public class PCentroCusto {
             retorno.setId(rs.getInt("id"));
             retorno.setDescricao(rs.getString("descricao"));
             retorno.setStatus(rs.getInt("status"));
-            
+
         }
         return retorno;
     }
 
     public List<CentroCusto> listar() throws SQLException {
 
-        String sql = "SELECT * FROM centro_custo";
+        String sql = "SELECT * FROM centro_custo ORDER BY id";
 
         Connection conexao = util.Conexao.getConexao();
         Statement st = conexao.createStatement();
@@ -105,12 +122,11 @@ public class PCentroCusto {
             cc.setId(rs.getInt("id"));
             cc.setDescricao(rs.getString("descricao"));
             cc.setStatus(rs.getInt("status"));
-      
+
             retorno.add(cc);
         }
 
         return retorno;
     }
-    
-    
+
 }
